init servant-based API for script exporting
